Initial commit for command zocalo.wrap
This is mostly a copy of dlstbx.wrap and still relies on some site
configuration files. Technically this should resolve #10.

zocalo.wrap differs to dlstbx.wrap in that it expects wrappers to
be declared in entry point 'zocalo.wrappers' instead of
'dlstbx.wrappers'. It uses the zocalo graylog configuration and,
at this time, relies on DLS-specific configuration files that are
not included in the package.