Drop plan to support --live, --test, --env-name
The alternative, '-e/--environment live' etc. has two benefits:
Firstly, the parameters won't clash with any other defined parameters,
as each command is in control of all names. Secondly, only a common
parameter prefix allows the user to specify the order in which
environments are loaded. Although this is not used right now it seems
kind of sensible to keep this option open for the future.