Add support for bindings to `RabbitMQAPI` (#147)
Bindings are handled slightly differently to queues and exchanges in the RabbitMQ HTTP API. A `POST` is required rather than a `PUT` to the address `api/bindings/e/{exchange name}/{e, q}/{queue or exchange name}`. The `routing_key` and `arguments` can be declared in the body and a "name" (called the `properties_key`) is then created for the binding from the `routing_key` and a hash of `arguments`. A single binding can then be found at `api/bindings/e/{exchange name}/{e, q}/{queue or exchange name}/{properties_key}`. The proposed API allows for deletion of a single binding if `properties_key` is specified or of all bindings between a given source and destination otherwise.

Resolves #145 

Co-authored-by: Richard Gildea <rjgildea@users.noreply.github.com>
Co-authored-by: Markus Gerstel <2102431+Anthchirp@users.noreply.github.com>