Fix for bindings API support (#149)
When constructing a url to point to a binding queues and exchanges are denoted by `q` and `e` respectively. However, when binding information is returned in response to a `GET` request the `destination_type` is labelled as either `queue` or `exchange`. The result from the `GET` request then needs to be mapped to `q` or `e` for consistency and to correctly fill the `destination_type` field in `BindingInfo`.