Fix header parsing in `dlq_reinject` for RabbitMQ #162
The format of messages from `dlq_purge` has all headers placed under the header key with no secondary headers structure which is currently assumed by `dlq_reinject`.

`PikaTransport` squashes what RabbitMQ calls `headers` and other properties that Zocalo considers to be headers into a single `header` which is what is written out by `dlq_purge`. When reinjecting the message we drop the properties not in the original RabbitMQ `headers` from the header before reinjecting so this should be fine and `dlq_reinject` should not look for a secondary `headers` structure within the header.