Regenerate SLURM models.py
It has proven hard to reproduce the exact contents of the (automatically
generated) models.py. There seems to be differences in flags between
what is the documented way it was run vs the contents, the SLURM version,
and the fact that datamodel-code-generator output order does not seem
to be stable.

Thus, this is now regenerated with:

    datamodel-code-generator 0.25.8
    ripgrep

And the command:

    curl    -H "X-SLURM-USER-NAME: gda2" \
            -H "X-SLURM-USER-TOKEN: $(cat /dls_sw/apps/zocalo/secrets/slurmrest-gda2.tkn)" \
            https://slurm-rest.diamond.ac.uk:8443/openapi/v3 \
        | rg --passthru -N '((?:db)?v)0.0.(\d+)_' -r '${1}00$2.' \
        | datamodel-codegen --target-python-version=3.8 \
            --use-schema-description \
            --use-field-description \
            --output-model-type=pydantic.BaseModel \
            --field-constraints \
            --use-double-quotes \
            --output models \
        && cp models/v0040.py src/zocalo/util/slurm/models.py