Switch pyproject backend to poetry
This makes it easier to consistently handle dependencies and dependency locks.

This should not affect any existing workflows, as pre-commit hooks to update
the lock and export to the requirements_dev.txt should make this mostly
transparent to any existing usage.

It means that it is easier to maintain consistency with the (fully specified
and verified) dependencies as they change.